// src/app/ipn/types/payment_status.type.ts
export enum PaymentStatus {
  APPROVED = 'APPROVED',
  AUTHORIZED = 'AUTHORIZED',
  PENDING = 'PENDING',
  REJECTED = 'REJECTED',
  REFUNDED = 'REFUNDED',
  CANCELLED = 'CANCELLED',
  CHARGEBACK = 'CHARGEBACK',
  IN_PROCESS = 'IN_PROCESS',
  IN_MEDIATION = 'IN_MEDIATION',
  REJECTED_BY_BANK = 'REJECTED_BY_BANK',
  ERROR = 'ERROR',
  UNKNOWN = 'UNKNOWN',
  WAITING_PAYMENT = 'WAITING_PAYMENT',
}